--- src/main/java/com/prgrms/devkdtorder/OrderRepository.java
@@ -1,7 +1,0 @@
-package com.prgrms.devkdtorder;
-
-public interface OrderRepository {
-
-    void insert(Order order) ;
-
-}
