+++ src/main/java/com/prgrms/devkdtorder/order/domain/Order.java
@@ -1,4 +1,6 @@
-package com.prgrms.devkdtorder;
+package com.prgrms.devkdtorder.order.domain;
+
+import com.prgrms.devkdtorder.voucher.domain.Voucher;
 
 import java.util.List;
 import java.util.Optional;
