+++ src/main/java/com/prgrms/devkdtorder/order/repository/OrderRepository.java
@@ -1,0 +1,9 @@
+package com.prgrms.devkdtorder.order.repository;
+
+import com.prgrms.devkdtorder.order.domain.Order;
+
+public interface OrderRepository {
+
+    void insert(Order order) ;
+
+}
