+++ src/main/java/com/prgrms/devkdtorder/order/service/OrderService.java
@@ -1,5 +1,10 @@
-package com.prgrms.devkdtorder;
+package com.prgrms.devkdtorder.order.service;
 
+import com.prgrms.devkdtorder.voucher.domain.Voucher;
+import com.prgrms.devkdtorder.voucher.service.VoucherService;
+import com.prgrms.devkdtorder.order.domain.Order;
+import com.prgrms.devkdtorder.order.domain.OrderItem;
+import com.prgrms.devkdtorder.order.repository.OrderRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
