+++ src/main/java/com/prgrms/devkdtorder/voucher/domain/FixedAmountVoucher.java
@@ -1,4 +1,6 @@
-package com.prgrms.devkdtorder;
+package com.prgrms.devkdtorder.voucher.domain;
+
+import com.prgrms.devkdtorder.voucher.domain.Voucher;
 
 import java.util.UUID;
