+++ src/main/java/com/prgrms/devkdtorder/voucher/domain/Voucher.java
@@ -1,4 +1,4 @@
-package com.prgrms.devkdtorder;
+package com.prgrms.devkdtorder.voucher.domain;
 
 import java.util.UUID;
