+++ src/main/java/com/prgrms/devkdtorder/voucher/repository/FileVoucherRepository.java
@@ -1,15 +1,14 @@
-package com.prgrms.devkdtorder;
+package com.prgrms.devkdtorder.voucher.repository;
 
+import com.prgrms.devkdtorder.voucher.domain.VoucherType;
+import com.prgrms.devkdtorder.voucher.domain.Voucher;
 import org.ini4j.Wini;
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Repository;
 
-import javax.swing.text.html.Option;
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
