+++ src/main/java/com/prgrms/devkdtorder/voucher/repository/MemVoucherRepository.java
@@ -1,5 +1,6 @@
-package com.prgrms.devkdtorder;
+package com.prgrms.devkdtorder.voucher.repository;
 
+import com.prgrms.devkdtorder.voucher.domain.Voucher;
 import org.springframework.stereotype.Repository;
 
 import java.util.*;
